Refactoring helper method from Util.cs to LocalsUtil.cs and refactoring code to look concise
