+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/Util.cs
@@ -1017,19 +1017,5 @@ EndProject");
             // Simply test the extension as that is all we care about
             return string.Equals(Path.GetExtension(configFileName), ".json", StringComparison.OrdinalIgnoreCase);
         }
-
-        public static string CreateDummyConfigFile(string directoryPath)
-        {
-            string[] lines = {"<?xml version=\"1.0\" encoding=\"utf-8\"?>",
-                              "<configuration>",
-                              "<config>",
-                              "<add key=\"foo\" value=\"bar\" />",
-                              "<add key=\"kung foo\" value=\"panda\" />",
-                              "</config>",
-                              "</configuration>" };
-            var dummyConfigPath = Path.Combine(directoryPath, @"NuGet.config");
-            File.WriteAllLines(dummyConfigPath, lines);
-            return dummyConfigPath;
-        }
     }
 }
